weather-chat/requirements.txt
streamlit>=1.34
google-cloud-aiplatform[preview]==1.56.0   #  ← 1.55.0 or newer
google-generativeai>=0.4.0                 # if you use it elsewhere
requests>=2.32